Revert "Revert "create custom hook for fetching products""
This reverts commit 46e54649e43013bee11539951f6ecb65683fff63.
