+++ locales/en/header.json
@@ -1,8 +1,1 @@
-{
-  "blogs": "Blogs",
-  "contact": "Contact",
-  "home": "Home",
-  "posts": "Posts",
-  "pricing": "Pricing",
-  "products": "Products"
-}
+{"blogs":"Blogs","contact":"Contact","home":"Home","posts":"Posts","products":"Products"}
