+++ locales/es/header.json
@@ -1,8 +1,1 @@
-{
-  "blogs": "Blogs",
-  "contact": "Contacto",
-  "home": "Hogar",
-  "posts": "Publicaciones",
-  "pricing": "Precios",
-  "products": "Productos"
-}
+{"blogs":"Blogs","contact":"Contacto","home":"Hogar","posts":"Publicaciones","products":"Productos"}
