+++ locales/ka/header.json
@@ -1,8 +1,1 @@
-{
-  "blogs": "ბლოგები",
-  "contact": "კონტაქტი",
-  "home": "მთავარი",
-  "posts": "პოსტები",
-  "pricing": "ფასები",
-  "products": "პროდუქტები"
-}
+{"blogs":"ბლოგები","contact":"კონტაქტი","home":"მთავარი","posts":"პოსტები","products":"პროდუქტები"}
