+++ src/app/[locale]/(dashboard)/posts/page.tsx
@@ -12,7 +12,6 @@ import { usePathname, useSearchParams } from "next/navigation";
 import "../../../../index.css";
 import "./index.css";
 import { deletePost } from "@/src/app/modules/deletePost";
-import { SetPosts } from "@/src/app/types/PostDetails";
 
 const PostsPage: React.FC<PostsAndProductPageType> = () => {
   const searchParams = useSearchParams();
